% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_helper.R
\name{addGeneName}
\alias{addGeneName}
\title{Add gene_id and gene_name}
\usage{
addGeneName(annotation, intron_database)
}
\arguments{
\item{annotation}{ENSEMBL annotation in granges}
}
\value{
trimmed exon
}
\description{
Add gene_id and gene_name
}
