% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_helper.R
\name{getExonStart}
\alias{getExonStart}
\title{Retrieve Exon 5SS first 25 nucleotides}
\usage{
getExonStart(x)
}
\arguments{
\item{grExon}{GenomicRanges exon}
}
\value{
trimmed exon
}
\description{
Retrieve Exon 5SS first 25 nucleotides
}
