% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_helper.R
\name{getIntronFraction}
\alias{getIntronFraction}
\title{Retrieve intron 3SS last 0.3 nucleotide fraction}
\usage{
getIntronFraction(grIntron)
}
\arguments{
\item{grIntron}{GenomicRanges intron}
}
\value{
trimmed intron
}
\description{
Retrieve intron 3SS last 0.3 nucleotide fraction
}
