% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-class.R
\docType{class}
\name{RegionsFlow-class}
\alias{RegionsFlow-class}
\alias{RegionsFlow}
\alias{FlowRegion}
\alias{FlowRegion<-,RegionsFlow-method}
\alias{'refintronFlow<-',RegionsFlow-method}
\alias{intronFlow<-}
\alias{intronFlow<-,RegionsFlow-method}
\alias{'intronFlow<-',RegionsFlow-method}
\alias{exonFlow<-}
\alias{metadata<-}
\alias{metadata}
\alias{metadata,RegionsFlow-method}
\alias{metadata,regionFlow-method}
\alias{quantFlow<-,RegionsFlow-method}
\alias{'quantFlow<-',RegionsFlow-method}
\alias{show,RegionsFlow-method}
\alias{show}
\title{S4 class for intron annotation data for a specific annotation version}
\usage{
RegionsFlow(
  FlowRegion = GenomicRanges::GRanges(),
  intronFlow = GenomicRanges::GRanges(),
  exonFlow = GenomicRanges::GRanges(),
  quantFlow = GenomicRanges::GRanges(),
  metadata = data.frame()
)

FlowRegion(x)

\S4method{FlowRegion}{RegionsFlow}(x) <- value

intronFlow(x) <- value

\S4method{intronFlow}{RegionsFlow}(x) <- value

exonFlow(x) <- value

metadata(x) <- value

metadata(x)

\S4method{metadata}{RegionsFlow}(x)

\S4method{quantFlow}{RegionsFlow}(x) <- value

\S4method{show}{RegionsFlow}(object)
}
\arguments{
\item{FlowRegion}{A GRanges object containing annotated intron ranges}

\item{intronFlow}{A data.frame containing mapping between transcript,
TSS, promoter and gene ids}

\item{exonFlow}{A GRanges object containing promoter coordinates}

\item{quantFlow}{A GRanges object containing promoter coordinates}

\item{metadata}{A GRanges object containing promoter coordinates}

\item{x}{A RegionsFlow object}
}
\value{
A promoter annotation object with three slots: intronRanges,
  promoterIdMapping and promoter Coordinates
}
\description{
S4 class for intron annotation data for a specific annotation version
}
\section{Functions}{
\itemize{
\item \code{FlowRegion()}: Getter for FlowRegion

\item \code{FlowRegion(RegionsFlow) <- value}: Setter for intronRanges

\item \code{intronFlow(x) <- value}: Setter for intronFlow

\item \code{intronFlow(RegionsFlow) <- value}: Setter for intronFlow

\item \code{exonFlow(x) <- value}: Setter for intronFlow

\item \code{metadata(x) <- value}: Setter for intronFlow

\item \code{metadata()}: Getter for metadata

\item \code{metadata(RegionsFlow)}: Getter for metadata

\item \code{quantFlow(RegionsFlow) <- value}: Setter for quantFlow

}}
\section{Slots}{

\describe{
\item{\code{FlowRegion}}{A GRanges object. The intron ranges annotated with the
promoter information.}

\item{\code{introns}}{A GRanges object. intron coordinates}

\item{\code{exonFlow}}{A GRanges object. exon coordinates}

\item{\code{quantFlow}}{A GRanges object. exon coordinates}

\item{\code{metadata}}{data.frame specifying the links between the two.
gene id and internal promoter state}
}}

