% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-introns.R
\name{retrieveIntrons}
\alias{retrieveIntrons}
\title{Retrieve introns from annotation}
\usage{
retrieveIntrons(refAnnotation.path, junctions)
}
\arguments{
\item{refAnnotation.path}{path to reference annotation in gtf format}
}
\value{
regionFlow-class
}
\description{
Retrieve introns from annotation
}
