% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice-flow.R
\name{SpliceFlow}
\alias{SpliceFlow}
\title{SpliceFlow calculate splicing efficiencies}
\usage{
SpliceFlow(se_counts, regionsFlow)
}
\arguments{
\item{se_counts}{: counts from exons and introns}

\item{regionsFlow:}{regionFlow-object containing exons and pairs}
}
\value{
SpliceFlow calculations
}
\description{
SpliceFlow calculate splicing efficiencies
}
